webapp/requirements.txt
css
Bootstrap v3.3.7



js
Bootstrap v3.3.7
jQuery v3.3.1 


python:
pandas
flask
mysql-connector-python v2.0.4